编码器获取速度转化为实际速度，优化冗余代码，优化调参界面，但屏幕调参仍然没搓出来

+++ Demo/CODE/action.h
@@ -1,5 +1,5 @@
 #include "headfile.h"
-extern float speed_goal;
+extern int speed_goal;
 void Track_Action(int* temp);
 void Obstacle_Action(void);
 //void big_left_circ()
