+++ Demo/CODE/speed.h
@@ -4,7 +4,7 @@
 void duty_set(int left_duty,int right_duty);
 void Speed_Loop();
 void Get_Actual_Speed();
-float speed_section(float error,char i);
+int speed_section(float error,char i);
 extern float error;
 extern int duty_L,duty_R;
 #endif
